+++ Snabble/UI/ShoppingCart/ShoppingCartViewController.swift
@@ -24,11 +24,6 @@ public protocol ShoppingCartDelegate: AnalyticsDelegate, MessageDelegate {
     /// - Parameter error: the error from the backend
     /// - Returns: true if the error has been dealt with and no error messages need to be shown from the SDK
     func handleCheckoutError(_ error: SnabbleError) -> Bool
-
-    /// called to get a customer's loyalty card number. if not nil, this number is passed to the backend as the
-    /// `loyaltyCard` property of the shopping cart, and it is also displayed as part of the QR code, if any of
-    /// the `encodedCodes` payment methods is used
-    func getLoyaltyCard(_ project: Project) -> String?
 }
 
 extension ShoppingCartDelegate {
@@ -38,10 +33,6 @@ extension ShoppingCartDelegate {
 
     public func handleCheckoutError(_ error: SnabbleError) -> Bool {
         return false
-    }
-
-    public func getLoyaltyCard(_ project: Project) -> String? {
-        return nil
     }
 }
 
@@ -311,10 +302,6 @@ public final class ShoppingCartViewController: UIViewController {
         spinner.centerYAnchor.constraint(equalTo: button.centerYAnchor).isActive = true
         button.isEnabled = false
 
-        if self.shoppingCart.loyaltyCard == nil {
-            self.shoppingCart.loyaltyCard = self.delegate.getLoyaltyCard(project)
-        }
-
         self.shoppingCart.createCheckoutInfo(SnabbleUI.project, timeout: 10) { result in
             spinner.stopAnimating()
             spinner.removeFromSuperview()
@@ -369,7 +356,7 @@ public final class ShoppingCartViewController: UIViewController {
         self.present(alert, animated: true)
     }
 
-    func updateTotals() {
+    public func updateTotals() {
         let count = self.shoppingCart.numberOfProducts
 
         self.tabBarItem.image = UIImage.fromBundle(count == 0 ? "icon-cart-inactive-empty" : "icon-cart-inactive-full")